refactor: ♻️ refactoring bus apis / util functions
